Moved `get_comet_experiment_url()` into utils.py
