+++ trl/trainer/bco_trainer.py
@@ -52,8 +52,8 @@ from transformers.trainer_utils import EvalLoopOutput, has_length
 from transformers.utils import is_peft_available
 
 from ..data_utils import maybe_apply_chat_template
-from ..integration_utils import get_comet_experiment_url
 from ..models import PreTrainedModelWrapper, create_reference_model
+from ..utils import get_comet_experiment_url
 from .bco_config import BCOConfig
 from .utils import (
     DPODataCollatorWithPadding,
