+++ trl/trainer/orpo_trainer.py
@@ -51,8 +51,8 @@ from transformers.utils import is_peft_available, is_torch_fx_proxy
 from transformers.utils.deprecation import deprecate_kwarg
 
 from ..data_utils import maybe_apply_chat_template, maybe_extract_prompt
-from ..integration_utils import get_comet_experiment_url
 from ..models import PreTrainedModelWrapper
+from ..utils import get_comet_experiment_url
 from .orpo_config import ORPOConfig
 from .utils import (
     DPODataCollatorWithPadding,
